% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week2_exercise.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Function fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number in USA.}

\item{year}{The year number}
}
\value{
This function returns a map of the specific state with all motor crash locations
        That happened in the specific year in that state. All data is taken from relevant file.
}
\description{
This function gets as input a state number (in USA) and a year.
Its output is a map of this state with points located at the geographical
location of motor crashes that happened in that state in that year.
Note that invalid year (where no file exists for that year), or invalid state number
will give an appropriate error.
}
\examples{
fars_map_state(1,2015)
fars_map_state(18,"2013")

}
