% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week2_exercise.R
\name{fars_read}
\alias{fars_read}
\title{Function fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the file name to read.
Either full path should be given, or path relative to local working directory.
If file name does not exist - an Error will be returned.}
}
\value{
This function returns a data frame object of the filename that was read.
}
\description{
This function gets as input a file name (of csv format), and
and reads it to a data frame object, which is what the function returns.
}
\examples{
fars_read("accident_2015.csv")
fars_read("C:\\\\coursera\\\\R_packages\\\\data\\\\accident_2015.csv")

}
