% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week2_exercise.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Function fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of the years to be read.}
}
\value{
This function returns a list of tables - one table per year.
         Each table has 2 columns: MONTH and year taken from original file of that year.
}
\description{
This function gets a list of years, and returns a list of tables for each year
in the input argument. the table for each year include the year and the months from the
relevant file of that year.
The function also checks for valid data files for those years.
A valid data file is a file which exists for this specific year.
}
