% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week2_exercise.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Function fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of the years to be read.}
}
\value{
This function returns data frame with a "MONTH" column,
and one columns per each year in the "years" input argument.
The values under each year column - is the number of motor crashes
that happeded in that year - for this specific month.
}
\description{
This function gets as input a list of years, and gives a summary of how many motor crashes
there were in each month of each year in USA (taken from relevant files for those years).
}
\examples{
fars_summarize_years(c("2014","2015"))
fars_summarize_years(c(2015,2013))

}
