% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week2_exercise.R
\name{make_filename}
\alias{make_filename}
\title{Function make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Any format of 4 digits year number.}
}
\value{
This function returns a string of the relevant file name fo the specific year.
}
\description{
This function gets as input a year number (in any format),
and outputs the relevant data file name for that year
}
