% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number representing a state in the FARS data}

\item{year}{A number representing the year for analysis}
}
\value{
This function returns a plot of the number of accidents
for the specified state.
}
\description{
This function plots the number of Accidents for a Perticular state.
This function returns an error message if either the \code{state.num} or
\code{year} do not exist in Dataset
}
\examples{
\dontrun{
fars_map_state('01',2014)
}
}
