% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map accidents for a state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number}

\item{year}{The year for which the accidents should be mapped}
}
\value{
The return value will be a map of accidents for a state and year.
If one provides an invalid state, the function will provide an error indicating
and invalid State number. Also, if no accidents were found for a state and year
combination, then the function will return a message indicating that no accidents can
be plotted.
}
\description{
\code{fars_map_state} will plot the accidents for a provided state and year
}
\examples{
fars_map_state(3, 2017)
}
