% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string.}
}
\value{
This function returns a DataFrame from reading \code{filename}.
}
\description{
Create a DataFrame from a File. If the File doesn't exist
the function will stop with an error message.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}
