% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A file or a file path pointing to a csv}
}
\value{
The function will use the provided file_path to load the csv in memory.
 If the file cannot be found, the function will provide an error message.
}
\description{
Read a csv file
}
\examples{
fars_read('data/accident_2013.csv.bz2')

}
