% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of numbers representing years for analysis}
}
\value{
This function returns a list of DataFrames with the
   length of the vector \code{years} and two columns: MONTH and year.
}
\description{
Read files into a list of DataFrames with two columns, MONTH and Year,
over the vector of years input. All values in the vector years need to be
convertable into an integer otherwise the make_filename function wll error.
}
\examples{
\dontrun{
fars_read_years(c(2014,2015))
}
}
