% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple csv files for specific years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years, where each year must be a valid integer year}
}
\value{
A list of dataframes for each year
}
\description{
\code{fars_read_years} accepts a vector of years. For each year, the functon invokes
\code{make_filename} that creates a file name based on a provided year and then reads
the contents of the csv using \code{fars_read}. The function will return a warning if
and invalid year is provided.
}
\examples{
fars_read_years(c(2016, 2017))
}
