% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of numbers shows years for analysis}
}
\value{
This function returns a DataFrame with Months as rows
   and Years as columns. Each combination contains the count of
   Accidents.
}
\description{
Summarizes the count of Accidents by Month and Year for the
input vector \code{years}.
}
\examples{
\dontrun{
fars_summarize_years(2012:2014)
}
}
