% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Total fatality statitics for each year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years}
}
\value{
A table of total accident counts broken by year and month.
}
\description{
\code{fars_summarize_years} will compute total fatality statistics for each
year and month. The function takes in a list of years for which statistics
are computed.
}
\examples{
fars_summarize_years(c(2016, 2017))
}
