% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A integer Representing the year}
}
\value{
This function returns a string according to \code{year}.
}
\description{
Select a filename according to the year. The Value of \code{year}
must be integer otherwise an error message will be in result.
}
\examples{
make_filename(2014)

}
