% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename for a specific year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a numeric value specifying the year}
}
\value{
A string representing a filename in the following format
 'accident_year.csv.bz2', where year will be replaced by the provided year
}
\description{
Create a filename for a specific year
}
\examples{
make_filename(1999)
}
